separated difficulty_level.py and colors.py added finishing touches to the code :))
