+++ game_logic/colors.py
@@ -1,14 +1,6 @@
-from enum import Enum
-
 WHITE = (255, 255, 255)
 BLACK = (0, 0, 0)
 RED = (255, 0, 0)
 GREEN = (0, 255, 0)
 BLUE = (0, 206, 209)
 ORANGE = (255, 140, 0)
-
-
-class Level(Enum):
-    EASY = 0
-    MEDIUM = 1
-    HARD = 2
