+++ game_logic/difficulty_level.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class Level(Enum):
+    EASY = 0
+    MEDIUM = 1
+    HARD = 2
